+++ CHANGELOG.md
@@ -33,6 +33,7 @@ All notable changes to this project will be documented in this file.
 ## [0.0.3] - 2022-06-7
 
 ### Notes
+Commit ID: [main 11d1d9c](https://github.com/bracesproul/earmark-backend/commit/11d1d9c)
 
 #### Added
 - Added /api/earmark/visualize for fetching data to populate charts on the /dashboard/visualize page
@@ -43,3 +44,17 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Updated the public_token_exchange to write the institution name to firebase
+
+
+## [0.0.4] - 2022-06-8
+
+### Notes
+Commit ID: [main ](https://github.com/bracesproul/earmark-backend/commit/)
+
+#### Added
+- Added /api/earmark/getDynamicTransactions for fetching data to populate dynamic transactions table on the /dashboard/[ins_id] page
+- Added exported function to firestore for fetching access token which corresponds to the institution id
+
+### Fixed
+
+### Changed